backend/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-RESTful==0.3.10
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3
